Pequenas mudanças lol
