+++ Cargo.toml
@@ -1,17 +1,13 @@
 [package]
-name = "gol"
+name = "game-of-life"
 version = "0.1.0"
 edition = "2021"
 
-[workspace]
-members = [ 
-  "configuration",
-  "game", 
-  "runner", 
-  "view", 
-  "websocket"
-]
 
-[[bin]]
-name = "main"
-path = "./runner/src/main.rs"
+[dependencies]
+clap = { version = "4.5.27", features = ["derive"] }
+crossterm = "0.28.1"
+rand = "0.9.0"
+ratatui = "0.29.0"
+tracing = "0.1.41"
+tracing-subscriber = "0.3"
