--- configuration/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "configuration"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-clap = { version = "4.5.20", features = ["derive"] }
-crossterm = "0.28.1"
-game = { path = "../game"}
