--- configuration/src/parser.rs
@@ -1,10 +1,0 @@
-use super::Mode;
-
-impl ToString for Mode {
-    fn to_string(&self) -> String {
-        match self {
-            Self::Test => String::from("test"),
-            Self::Random => String::from("random"),
-        }
-    }
-}
