--- game/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "game"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-rand = "0.8.5"
-
