--- runner/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "runner"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-game = { path = "../game" }
-view = { path = "../view" }
-configuration = { path = "../configuration" }
-rand = "0.8.5"
