--- runner/src/main.rs
@@ -1,8 +1,0 @@
-use configuration::Config;
-use runner::Runner;
-
-fn main() {
-    let config = Config::configure();
-    let mut runner = Runner::new(config);
-    runner.run();
-}
