+++ src/configuration/mod.rs
@@ -1,0 +1,6 @@
+mod cli;
+mod internal_configuration;
+mod mode;
+
+pub use internal_configuration::Configuration;
+pub use mode::Mode;
