+++ src/configuration/mode.rs
@@ -1,0 +1,18 @@
+use std::fmt::Display;
+
+use clap::ValueEnum;
+
+#[derive(ValueEnum, Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+pub enum Mode {
+    Test,
+    Random,
+}
+
+impl Display for Mode {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Mode::Test => write!(f, "test"),
+            Mode::Random => write!(f, "random"),
+        }
+    }
+}
