+++ src/conway/mod.rs
@@ -1,0 +1,13 @@
+mod cell;
+mod game;
+mod matrix;
+mod neighbour;
+mod point;
+mod state;
+
+pub use cell::Cell;
+pub use game::ConwayGame;
+pub use matrix::Matrix;
+pub use neighbour::Neighbour;
+pub use point::Point;
+pub use state::State;
