+++ src/conway/state.rs
@@ -1,0 +1,5 @@
+#[derive(PartialEq, Clone, Copy, Debug, Eq)]
+pub enum State {
+    Alive,
+    Dead,
+}
