+++ src/lib.rs
@@ -1,0 +1,10 @@
+mod configuration;
+mod conway;
+mod network;
+mod runner;
+mod utils;
+mod view;
+
+pub use configuration::Configuration;
+pub use conway::ConwayGame;
+pub use runner::Runner;
