--- view/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "view"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-crossterm = "0.28.1"
-ratatui = "0.29.0"
-game = { path = "../game"}
