--- websocket/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "websocket_game"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-async-trait = "0.1.52"
-ezsockets = { version="*", features = ["tungstenite"] }
-tokio = { version = "1.17.0", features = ["full"] }
-tracing = "0.1.32"
-tracing-subscriber = "0.3.9"
-
-
-[[bin]]
-name = "client"
-path = "src/client.rs"
-
-[[bin]]
-name = "server"
-path = "src/server.rs"
-
-
